backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
groq>=0.9.0
psycopg2-binary>=2.9.0
